+++ src/spatialdata/transformations/ngff/ngff_coordinate_system.py
